add support for exact patient searches
Modified the get API for patient object to support a
new exact query param. If "true", then last name and
first name will be checked for a case insensitive
exact match (e.g., paternal_last__iexact in Django-speak).
This is useful when we go to search the database for an
existing patient before deciding if we are doing a create
or update. A search like paternal_last__icontains will
cause a false match in these cases if a substring is used,
e.g., "go" will match "gonzalez" and when we are searching
for a patient with last name "gomez", hence a false result.