Fix patient dup verification returns patient 1
We've had a long standing bug where the check for
duplicate patient logic returns patient 1 maybe a
small number of times during a clinic.

Logs collected in commit 034e661 indicated the problem
was happening when django issued 303 redirects for patient
urls but only when a spanish language, non-ascii char
was in the name(s) being searched. This redirect was
happening automatically due to django configuration for
URLs that didn't end with '/', which is the case with
the URLs emitted by the android apps. In Django, this
is controlled by the APPEND_SLASH settings.

In this PR, we change just a couple of these, the patient
and createuser URLs since both are the most likely to
see non-ascii. Future PR will add more to avoid all the
303s since there is probably a performance hit for django
reissuing them (and it doubles by definition the size of
the apache2 http access logs). We also will make changes
in the next update to the android apps to fix the issue
at the origin.

In any event, the redirect would be a bare patient URL
without any query args. This GET URL would then return
a list of ALL patient IDs, sorted, from which we chose
the first response in the android apps, which was
patient #1, This is why we would see patient 1 as the
reported duplicate.